requirements.txt
gdown
datasets
python-mecab-ko